fix: correct Shi-Tomasi min-eigenvalue formula
compute_min_eigenvalues_into placed sqrt() on the wrong subexpression:

    Before: sqrt(trace - discriminant) / 2
    After:  (trace - sqrt(discriminant)) / 2

The closed-form min-eigenvalue of [[a, c], [c, b]] is
(trace - sqrt((a-b)^2 + 4 c^2)) / 2. With sqrt() on the full
trace-discriminant expression, the argument was overwhelmingly negative
on natural imagery and most pixels yielded NaN. Downstream NaN handling
(partial_cmp().unwrap_or(Equal) sort, max fold, >= retain) produced
near-arbitrary detection output. good_features_to_track and
OpticalFlowBuffer::reset_with_good_features_to_track now return
meaningful corner scores on natural inputs.

+++ src/features.rs
@@ -145,7 +145,7 @@ fn compute_min_eigenvalues_into(
 
             let trace = a_val + b_val;
             let discriminant = (a_val - b_val).pow(2) + 4 * c_val.pow(2);
-            let min_eigen = (((trace - discriminant) as f32).sqrt()) / 2.0;
+            let min_eigen = (trace as f32 - (discriminant as f32).sqrt()) / 2.0;
 
             out.push((x, y, min_eigen));
         }
